+++ semana8/semana8-projeto/src/components/ButtonsMatch/ButtonsMatch.js
@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from 'react';
+
 import {ContainerButtons, ButtonDislike, ButtonLike} from "./styles";
