+++ semana8/semana8-projeto/src/components/Matchs/Matchs.js
@@ -1,45 +1,11 @@
 import React, { useState, useEffect } from 'react';
-import './Initial.css';
-import axios from 'axios';
+
 import ProfileMatch from '../ProfileMatch/ProfileMatch';
-import styled from 'styled-components';
 
-const ContainerMatches = styled.div `
-  width: 398px;
-  height: 589px;
-  display:flex;
-  flex-direction: column;
-  justify-content:space-between;
-  overflow:auto;
-  
-`;
+import axios from 'axios';
 
-const ClearButton = styled.button `
-  width: 150px;
-  bottom: 0;
-  align-self: center;
-  margin-bottom: 10px;
-  padding: 5px 0;
-  border: 1px solid #51A89D;
-  color: #762D93;
-  font-weight: bold;
-  background-color: #fff;
-  transition: 200ms;
-  cursor:pointer;
-    :hover {
-      background-color: #51A89D;
-      color: #fff;
-      transform: scale(1.03)
-    }
-    :focus {
-      outline: none;
-    }
-`;
-
-const TitleMatch = styled.h2 `
-  color: #762D93;
-  text-align: center;
-`;
+import './Initial.css';
+import { ContainerMatches, ClearButton, TitleMatch } from './styled'
 
 function Matchs() {
   const [matchs, setMatchs] = useState([]);
